finalize function
