+++ src/pages/PartnerDetail.jsx
@@ -1,22 +1,69 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import { Container, Row, Col } from "react-bootstrap";
 import { useParams, useNavigate, Link } from "react-router-dom";
-import { FiX } from "react-icons/fi";
-import PartnerList from "../components/PartnerList";
+import { FiArrowLeft } from "react-icons/fi";
 import JoinUsButton from "../components/JoinUsButton";
-import "../css/PartnerDetail.css"; // 可选：新增页面专属样式
+import axios from "axios";
+import { FaUserPlus } from "react-icons/fa";
+import "../css/PartnerDetail.css";
+
+const productTitleMap = {
+  Studyfin: "留学中介",
+  "罗塞尼斯半岛度假村": "旅游中介",
+  "AI美甲": "加盟商",
+};
+
+function getMediaUrl(media) {
+  if (!media) return null;
+  if (Array.isArray(media)) return getMediaUrl(media[0]);
+  if (media.url) return import.meta.env.VITE_STRAPI_HOST + media.url;
+  if (media.data?.attributes?.url) return import.meta.env.VITE_STRAPI_HOST + media.data.attributes.url;
+  return null;
+}
 
 const PartnerDetail = () => {
   const { productName } = useParams();
   const decodedProductName = decodeURIComponent(productName);
   const navigate = useNavigate();
 
+  const [partners, setPartners] = useState([]);
+  const [documentId, setDocumentId] = useState("");
+  const [showAll, setShowAll] = useState(false);
+
+  useEffect(() => {
+    const fetchPartners = async () => {
+      try {
+        const params = new URLSearchParams();
+        params.append("filters[productName][$eq]", decodedProductName);
+        params.append("populate", "*");
+
+        const url = `${import.meta.env.VITE_STRAPI_HOST}/api/partner-application-submissions?${params.toString()}`;
+        const res = await axios.get(url, {
+          headers: { Authorization: `Bearer ${import.meta.env.VITE_API_KEY_MERCHANT_UPLOAD}` },
+        });
+
+        const entries = res.data?.data || [];
+        setPartners(entries);
+        setDocumentId(entries[0]?.documentId || "");
+
+      } catch (err) {
+        console.error("拉取合作伙伴失败", err);
+        setPartners([]);
+        setDocumentId("");
+      }
+    };
+
+    fetchPartners();
+  }, [decodedProductName]);
+
+  const approvedPartners = partners.filter(p => (p.attributes || p).approved);
+  const visiblePartners = showAll ? approvedPartners : approvedPartners.slice(0, 2);
+  const title = productTitleMap[decodedProductName] || "合作伙伴";
+
   return (
     <Container style={{ paddingTop: "80px", paddingBottom: "40px", position: "relative" }}>
-
-      {/* 右上角 X 关闭按钮 */}
       <div
-        onClick={() => navigate(-1)}
+        onClick={() => navigate(`/products/${encodeURIComponent(decodedProductName)}`)}
         style={{
           position: "absolute",
           top: "20px",
@@ -25,15 +72,75 @@ const PartnerDetail = () => {
           fontSize: "24px",
           color: "#555",
         }}
-        title="关闭"
+        title="返回"
       >
-        <FiX />
+        <FiArrowLeft />
       </div>
 
-      {/* 合作伙伴列表 */}
-      <PartnerList currentProductName={decodedProductName} />
+      <h5 className="partner-section-title">{title}</h5>
 
-      {/* 加入按钮 */}
+      {approvedPartners.length === 0 ? (
+        <p>敬请期待</p>
+      ) : (
+        <>
+          <div className="partner-list-container">
+            {visiblePartners.map((item, idx) => {
+              const attr = item.attributes || item || {};
+              const avatarUrl = getMediaUrl(attr.advisorAvatar);
+              const asicUrl = getMediaUrl(attr.asicCertificate);
+              const licenseUrl = getMediaUrl(attr.licenseFile);
+
+              return (
+                <div key={item.id || idx} className="partner-card">
+                  <div className="partner-logo-wrapper">
+                    {avatarUrl && <img src={avatarUrl} alt="顾问头像" className="partner-logo" />}
+                  </div>
+
+                  <div className="partner-main-info">
+                    <div className="info-section">
+                      <div className="info-section-title">💼 专业资质</div>
+                      <div className="partner-field">
+                        <span className="field-label">从业经验：</span>
+                        {attr.experienceYears || "未填写"}
+                      </div>
+                      {asicUrl && (
+                        <div className="partner-field">
+                          <span className="field-label">ASIC 证书：</span>
+                          <a href={asicUrl} target="_blank" rel="noopener noreferrer">📄 查看证书</a>
+                        </div>
+                      )}
+                      {licenseUrl && (
+                        <div className="partner-field">
+                          <span className="field-label">牌照文件：</span>
+                          <a href={licenseUrl} target="_blank" rel="noopener noreferrer">📎 查看牌照</a>
+                        </div>
+                      )}
+                    </div>
+                  </div>
+
+                  <div className="partner-join-button">
+                    <Link to={`/products/${encodeURIComponent(decodedProductName)}/CustomerApplicationForm?partnerID=${encodeURIComponent(attr.partnerID)}&documentId=${encodeURIComponent(documentId)}`}>
+                      <button className="custom-join-button">
+                        <FaUserPlus style={{ marginRight: "6px" }} />
+                        立即咨询
+                      </button>
+                    </Link>
+                  </div>
+                </div>
+              );
+            })}
+          </div>
+
+          {approvedPartners.length > 2 && (
+            <div className="toggle-button-wrapper">
+              <button className="custom-join-button" onClick={() => setShowAll(!showAll)}>
+                {showAll ? "收起" : "显示全部"}
+              </button>
+            </div>
+          )}
+        </>
+      )}
+
       <Row className="mt-4 justify-content-start">
         <Col xs="auto">
           <Link to={`/products/${encodeURIComponent(decodedProductName)}/PartnerDetail/PartnerApplicationForm`}>
